adding the yorozuya-test module and configuring the Jacoco plugin to collecting coverage reports update README.md to add building badge correcting the project naming error
