+++ README.md
@@ -1,14 +1,20 @@
 # YOROZUYA
 
+![yorozuya](https://github.com/ksewen/yorozuya/actions/workflows/ci.yml/badge.svg)
 ![License](https://img.shields.io/badge/License-MIT-blue.svg)
 ![Java](https://img.shields.io/badge/Java-17-blue.svg)
 
 Dedicated to developing a ready-to-use enterprise-level microservice framework based on Spring Cloud.
 
-**YOROZUYA** is an organization in the Japanese manga "Gintama" and its derivative works, which takes on various tasks and requests from others.
-> When the pair rescues a teenage alien girl with super-strength, Kagura, from a Yakuza group, they accept her into their freelancing business and the three become known as "Yorozuya" (万事屋, "We do everything" or literally "The Anything Store").
+**YOROZUYA** is an organization in the Japanese manga "Gintama" and its derivative works, which takes on various tasks
+and requests from others.
+> When the pair rescues a teenage alien girl with super-strength, Kagura, from a Yakuza group, they accept her into
+> their freelancing business and the three become known as "Yorozuya" (万事屋, "We do everything" or literally "The
+> Anything Store").
 
 ## Trademarks and licenses
+
 The source code is licensed under [MIT license](https://opensource.org/license/mit/)
 
-Spring, Spring Boot and Spring Cloud are trademarks of [Pivotal Software, Inc.](https://tanzu.vmware.com/) in the U.S. and other countries.
+Spring, Spring Boot and Spring Cloud are trademarks of [Pivotal Software, Inc.](https://tanzu.vmware.com/) in the U.S.
+and other countries.
