+++ pom.xml
@@ -20,10 +20,10 @@
         <springdoc.version>2.2.0</springdoc.version>
         <yorozuya.version>${revision}</yorozuya.version>
         <guava.version>32.1.2-jre</guava.version>
-        <jacoco.maven.plugin.version>0.8.10</jacoco.maven.plugin.version>
         <ok.http.3.version>4.11.0</ok.http.3.version>
         <http.client.version>5.2.1</http.client.version>
 
+        <maven.jacoco.plugin.version>0.8.10</maven.jacoco.plugin.version>
         <maven.flatten.plugin.version>1.5.0</maven.flatten.plugin.version>
         <maven.source.plugin.version>3.3.0</maven.source.plugin.version>
     </properties>
@@ -50,6 +50,11 @@
             </dependency>
 
             <!-- yorozuya -->
+            <dependency>
+                <groupId>com.github.ksewen</groupId>
+                <artifactId>yorozuya-common</artifactId>
+                <version>${yorozuya.version}</version>
+            </dependency>
             <dependency>
                 <groupId>com.github.ksewen</groupId>
                 <artifactId>yorozuya-spring-boot-starter</artifactId>
@@ -93,7 +98,7 @@
             <modules>
                 <module>yorozuya-common</module>
                 <module>yorozuya-spring-boot-starter</module>
-                <module>yoruzaya-samples</module>
+                <module>yorozuya-samples</module>
             </modules>
         </profile>
         <profile>
@@ -103,10 +108,27 @@
                 <module>yorozuya-spring-boot-starter</module>
             </modules>
         </profile>
+        <profile>
+            <id>test</id>
+            <modules>
+                <module>yorozuya-common</module>
+                <module>yorozuya-spring-boot-starter</module>
+                <module>yorozuya-test</module>
+            </modules>
+        </profile>
 
     </profiles>
 
     <build>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>org.jacoco</groupId>
+                    <artifactId>jacoco-maven-plugin</artifactId>
+                    <version>${maven.jacoco.plugin.version}</version>
+                </plugin>
+            </plugins>
+        </pluginManagement>
         <plugins>
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
