+++ yorozuya-samples/rest-client-httpclient/src/main/resources/application.yml
