+++ yorozuya-samples/rest-client-httpclient/src/test/java/com/github/ksewen/yorozuya/sample/rest/client/httpclient/RestClientOkhttpApplicationTests.java
@@ -1,0 +1,11 @@
+package com.github.ksewen.yorozuya.sample.rest.client.httpclient;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RestClientOkhttpApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
