+++ yorozuya-samples/rest-client-okhttp/src/main/java/com/github/ksewen/yorozuya/sample/rest/client/okhttp/RestClientOkhttpApplication.java
@@ -1,0 +1,12 @@
+package com.github.ksewen.yorozuya.sample.rest.client.okhttp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RestClientOkhttpApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(RestClientOkhttpApplication.class, args);
+  }
+}
