+++ yorozuya-samples/rest-client-okhttp/src/main/resources/application.yml
