+++ yorozuya-samples/rest-client-okhttp/src/test/java/com/github/ksewen/yorozuya/sample/rest/client/okhttp/RestClientOkhttpApplicationTests.java
@@ -1,0 +1,11 @@
+package com.github.ksewen.yorozuya.sample.rest.client.okhttp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RestClientOkhttpApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
