+++ yorozuya-test/pom.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.github.ksewen</groupId>
+        <artifactId>yorozuya</artifactId>
+        <version>${revision}</version>
+    </parent>
+    <artifactId>yorozuya-test</artifactId>
+    <version>${revision}</version>
+    <name>yorozuya-test</name>
+    <description>yorozuya-test</description>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.github.ksewen</groupId>
+            <artifactId>yorozuya-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.github.ksewen</groupId>
+            <artifactId>yorozuya-spring-boot-starter</artifactId>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.jacoco</groupId>
+                <artifactId>jacoco-maven-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>report-aggregate</id>
+                        <phase>test</phase>
+                        <goals>
+                            <goal>report-aggregate</goal>
+                        </goals>
+                        <configuration>
+                            <title>YOROZUYA</title>
+                            <footer>Code Coverage Report for YOROZUYA ${revision}</footer>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
