--- yoruzaya-samples/rest-client-httpclient/src/test/java/com/github/ksewen/rest/client/httpclient/RestClientOkhttpApplicationTests.java
@@ -1,11 +1,0 @@
-package com.github.ksewen.rest.client.httpclient;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestClientOkhttpApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
