--- yoruzaya-samples/rest-client-okhttp/src/main/java/com/github/ksewen/rest/client/okhttp/RestClientOkhttpApplication.java
@@ -1,12 +1,0 @@
-package com.github.ksewen.rest.client.okhttp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RestClientOkhttpApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(RestClientOkhttpApplication.class, args);
-  }
-}
