--- yoruzaya-samples/rest-client-okhttp/src/test/java/com/github/ksewen/rest/client/okhttp/RestClientOkhttpApplicationTests.java
@@ -1,11 +1,0 @@
-package com.github.ksewen.rest.client.okhttp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestClientOkhttpApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
